Move JTextAreaOptionPane to tools.image
tools.image is the only package that uses the JTextAreaOptionPane class.
The visibility of all JTextAreaOptionPane members was subsequently
reduced to package-private.
